requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.1
Flask-Migrate==4.0.4
PyMySQL==1.0.2
mysql-connector-python==9.0.0